Patch Like Functionality In Reducer
Previous version of the like logic in reducer had looked for a matching transaction object within the currentUser's sent_transactions, received_transactions, and the selectedProfile's sent_transactions and received_transactions.  Because the logic includes the use of Array.indexOf to find the transaction object's index, for the case statements where the transaction object did not exist, it stored '-1' in the index variable which would be used to as a part of an Array.splice.  This ultimately caused a newly liked transaction object to appear in multiple areas where it did not belong.